requirements.txt
anyio==3.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.1
fonttools==4.39.4
h11==0.14.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing==3.7.4.3
typing_extensions==4.5.0
uvicorn==0.22.0
